Move compiler phase selection to left-hand bar.
Ant has a `Tree` component that we can use to simplify the UI a bit. Rather than have a selection box above the graph, we can select the phase directly from the same UI element for selecting which method to work with. The change simplifies state management a fair bit, too. The tree nodes can store state about what phase has been selected and we've eliminated the state where a method is selected but a phase is not.